src/lib/fastapi/requirements.txt
fastapi
uvicorn
openai-whisper
pydantic
